Add CLA checking

+++ lib/lita/handlers/jls.rb
@@ -2,6 +2,7 @@ require "cabin"
 require "tmpdir"
 require "fileutils"
 require "insist"
+require "uri"
 
 module Lita
   module Handlers
@@ -9,15 +10,27 @@ module Lita
       require "lita-jls/util"
       include LitaJLS::Util
 
-      #route /^merge(?<dry>\?)? (?:(?<user>[A-Za-z0-9_-]+)\/)?(?<project>[A-Za-z0-9_-]+)#(?<pr>\d+) (?<branchspec>.*)$/, :merge,
       route /^merge(?<dry>\?)? (?<pr_url>[^ ]+) (?<branchspec>.*)$/, :merge,
+        :command => true,
         :help => { "merge https://github.com/ORG/PROJECT/pull/NUMBER branch1 [branch2 ...]" => "merges a PR into one or more branches. To see if a merge is successful, use 'merge? project#pr branch1 [branch2 ...]" }
-      route /^\(tableflip\)$/, :tableflip, :command => true, :help => { "(tableflip)" => "Fix whatever just broke. Probably git is going funky, so I will purge my local git junk" }
+
+      route /^cla(?<dry>\?)? (?<pr_url>[^ ]+)$/, :cla,
+        :command => true,
+        :help => { "cla https://github.com/ORG/PROJECT/pull/NUMBER" => "CLA check for a giaven PR" }
+
+      route /^\(tableflip\)$/, :tableflip,
+        :command => true,
+        :help => { "(tableflip)" => "Fix whatever just broke. Probably git is going funky, so I will purge my local git junk" }
 
       REMOTE = "origin"
       URLBASE = "https://github.com/"
 
       on :loaded, :setup
+      on :connected, :connected
+
+      def connected(*args)
+        p "OK"
+      end
 
       def self.default_config(config)
         config.default_organization = nil
@@ -28,12 +41,20 @@ module Lita
         @@logger_subscription ||= logger.subscribe(STDOUT)
       end
 
+      def cla(msg)
+        @cla_uri = config.cla_uri
+        pull = msg.match_data["pr_url"]
+        pull_path = URI.parse(pull).path
+        _, user, project, _, pr = pull_path.split("/")
+        cla?("#{user}/#{project}", pr)
+        msg.reply("#{user}/#{project}##{pr} CLA OK (freddie)")
+      rescue => e
+        msg.reply(e)
+      end
+
       def merge(msg)
+        @cla_uri = config.cla_uri
         FileUtils.mkdir_p(workdir) unless File.directory?(workdir)
-        #user = msg.match_data["user"] || "elasticsearch"
-        #project = msg.match_data["project"]
-        #pr = msg.match_data["pr"]
-        require "uri"
         pull = msg.match_data["pr_url"]
         pull_path = URI.parse(pull).path
         _, user, project, _, pr = pull_path.split("/")
@@ -44,7 +65,13 @@ module Lita
 
         branchspec = msg.match_data["branchspec"]
         dry_run = msg.match_data["dry"]
-        p :dry? => dry_run
+
+        begin
+          cla?("#{user}/#{project}", pr)
+        rescue => e
+          msg.reply("(firstworldproblems) cla check failed for #{user}/#{project}##{pr}.\n #{e}")
+          return
+        end
 
         url = File.join(URLBASE, user, project)
         #git_url = "git@github.com:/#{user}/#{project}.git"
@@ -57,8 +84,6 @@ module Lita
         repo = clone_at(git_url, gitpath)
 
         git(gitpath, "am", "--abort") if File.directory?(".git/rebase-apply")
-        git(gitpath, "reset", "--hard", "#{REMOTE}/master")
-        git(gitpath, "clean", "-f")
 
         # TODO(sissel): Fetch the PR patch
         logger.info("Fetching PR patch", :url => pr_url)
@@ -66,7 +91,7 @@ module Lita
         response = http.get(URI.parse(pr_url).path)
         if !response.success?
           logger.warn("Failed fetching patch", :url => pr_url, :status => response.status, :headers => response.headers)
-          msg.reply("Failed fetching patch. Cannot continue!")
+          msg.reply("(grumpycat) Failed fetching patch. Cannot continue!")
           return
         end
 
@@ -79,9 +104,8 @@ module Lita
             logger.info("Switching branches", :branch => branch, :repo => gitpath)
             #repo.checkout(branch)
             git(gitpath, "checkout", branch)
+            git(gitpath, "reset", "--hard", "#{REMOTE}/#{branch}")
             git(gitpath, "pull", "--ff-only")
-            #repo.reset("#{REMOTE}/#{branch}", :mixed)
-            git(gitpath, "reset", "#{REMOTE}/#{branch}")
             apply_patch(repo, patch) do |commit|
               # Append the PR number to commit message
 
@@ -90,7 +114,7 @@ module Lita
               commit[:message] += "\nFixes ##{pr}"
             end
           rescue => e
-            msg.reply("Failed attempting to merge #{user}/#{project}##{pr}: #{e}")
+            msg.reply("(jackie) Failed attempting to merge #{user}/#{project}##{pr} into #{branch}: #{e}")
             raise
           end
         end
@@ -106,7 +130,7 @@ module Lita
           github_issue_label("#{user}/#{project}", pr.to_i, labels)
         end
       rescue => e
-        msg.reply("Error: #{e.inspect}")
+        msg.reply("(stare) Error: #{e.inspect}")
         raise
       end # def merge
 
@@ -114,10 +138,10 @@ module Lita
         begin
           dir = workdir("gitbase")
           insist { dir } =~ /\/lita-jls/ # Just in case, before we go purging things...
-          FileUtils.rm_r(dir)
-          msg.reply("git (tableflip) (success)")
+          FileUtils.rm_r(dir) if File.directory?(dir)
+          msg.reply("Git: (tableflip) (success)")
         rescue => e
-          msg.reply("Git (tableflip) (huh): #{e}")
+          msg.reply("Git: (tableflip) (huh): #{e}")
           raise e
         end
       end
